requirements.txt
fastapi
uvicorn
pydantic
statsbombpy
pandas
numpy
streamlit
langchain
langchain-google-genai
google-generativeai
pyyaml
python-dotenv
langchain-openai